changes: move modules to infrastructure to modules layer
