+++ backend/src/infrastructure/modules/auth.module.ts
@@ -3,14 +3,14 @@ import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 import { PassportModule } from '@nestjs/passport';
 
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { UserRepository } from '../user/user.repository';
-import { TokenVersionModule } from '../token-version/token-version.module';
-import { AuthRepository } from './auth.repository';
-import { LocalStrategy } from './strategy/local.strategy';
-import { GoogleStrategy } from './strategy/google.strategy';
-import { JwtStrategy } from './strategy/jwt.strategy';
+import { TokenVersionModule } from './token-version.module';
+import { AuthController } from '../controllers/auth.controller';
+import { AuthRepository } from '../repositories/auth.repository';
+import { UserRepository } from '../repositories/user.repository';
+import { AuthService } from '@application/auth/use-case/auth.service';
+import { JwtStrategy } from '@application/auth/strategies/jwt.strategy';
+import { LocalStrategy } from '@application/auth/strategies/local.strategy';
+import { GoogleStrategy } from '@application/auth/strategies/google.strategy';
 
 @Module({
   imports: [
